refactor(server): isolate rls request context
